rings/requirements.txt
argcomplete==2.0.0
asgiref==3.4.1
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dataclasses==0.8
Django==3.2.12
django-cors-headers==3.10.1
django-model-utils==4.2.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
importlib-metadata==4.8.3
jmespath==0.10.0
kappa==0.6.0
placebo==0.9.0
psycopg2==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
troposphere==3.1.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wsgi-request-logger==0.4.6
zappa==0.54.1
zipp==3.6.0
